+++ CHANGES.md
@@ -27,6 +27,10 @@ Gererar portal para consulta de clientes/servicios dc basados en una base access
         N[pages/modals/sdc_abmhousing.php]
         N[pages/modals/sdc_abmhousing.js]
         N[pages/helpers/sdc_abmhousingdb.php]
+        M[pages/sdc_housing.php]
+    - Importación masiva de Servicios de Hosting via csv
+        M[pages/sdc_hosting.php]
+        bd: sdc_hosting_temp